Explicitly group aggregate measurements.
Previously, aggregate measurements were not explicitly grouped.  This
changes aggregate measurements to be a single row in the database, with
values a JSON document (which should be of type `{[aggregateType:
string]: number}`, enforced implicitly through the message schema).

Provides a migration to update the database schema and collapse existing
aggregate measurement groups into one row.